requirements.txt
pip~=25.0.1
attrs~=25.1.0
distro~=1.9.0
filelock~=3.17.0
Django~=5.1.7
djangorestframework~=3.15.2
asgiref~=3.8.1
sqlparse~=0.5.3
psycopg2~=2.9.10
psycopg2-binary~=2.9.10
colorama~=0.4.6
requests~=2.32.3
pytz~=2025.2
PyYAML~=6.0.2
packaging~=24.2
inflection~=0.5.1
uritemplate~=4.1.1
PyJWT~=2.9.0
djangorestframework-simplejwt~=5.4.0
pillow~=10.0.0
drf-yasg~=1.21.10
django-filter~=23.3